Fixes, closes #19709

+++ app-sci/foldingathome/files/folding-init.d
@@ -1,7 +1,7 @@
 #!/sbin/runscript
 # Copyright 1999-2003 Gentoo Technologies, Inc.
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/app-sci/foldingathome/files/folding-init.d,v 1.2 2003/02/27 10:23:31 aliz Exp $
+# $Header: /var/cvsroot/gentoo-x86/app-sci/foldingathome/files/folding-init.d,v 1.3 2003/04/23 08:20:25 aliz Exp $
 
 start() {
 
@@ -13,7 +13,9 @@ start() {
 
 stop() {
 	ebegin "Stopping Folding@home"
-	killall foldingathome
+	killall FahCore_78.exe > /dev/null
+	if ! $? ; then
+		killall foldingathome
+	fi
 	eend $?
 }
-
